Modify contraception services
