+++ application/service_sync/data_processing/tests/test_validation.py
@@ -1,12 +1,17 @@
-from unittest.mock import MagicMock, patch
+from unittest.mock import MagicMock, call, patch
 
 import pytest
 from aws_lambda_powertools.logging import Logger
 
+from application.common.constants import DOS_PALLIATIVE_CARE_SYMPTOM_DISCRIMINATOR, DOS_PALLIATIVE_CARE_SYMPTOM_GROUP
 from application.common.nhs import NHSEntity
-from application.service_sync.validation import validate_opening_times, validate_website
+from application.service_sync.data_processing.validation import (
+    validate_opening_times,
+    validate_website,
+    validate_z_code_exists,
+)
 
-FILE_PATH = "application.service_sync.validation"
+FILE_PATH = "application.service_sync.data_processing.validation"
 
 
 @patch(f"{FILE_PATH}.log_service_with_generic_bank_holiday")
@@ -83,3 +88,65 @@ def test_validate_website_failure(mock_log_website_is_invalid: MagicMock, websit
     # Act & Assert
     assert False is validate_website(nhs_entity=nhs_entity, nhs_website=website, dos_service=dos_service)
     mock_log_website_is_invalid.assert_called_once_with(nhs_entity, website, dos_service)
+
+
+@patch(f"{FILE_PATH}.query_dos_db")
+def test_validate_z_code_exists(mock_query_dos_db: MagicMock):
+    # Arrange
+    mock_connection = MagicMock()
+    mock_query_dos_db.return_value.rowcount = 1
+    dos_service = MagicMock()
+    # Act
+    response = validate_z_code_exists(
+        connection=mock_connection,
+        dos_service=dos_service,
+        symptom_group_id=DOS_PALLIATIVE_CARE_SYMPTOM_GROUP,
+        symptom_discriminator_id=DOS_PALLIATIVE_CARE_SYMPTOM_DISCRIMINATOR,
+        z_code_alias="Palliative Care",
+    )
+    # Assert
+    assert True is response
+    mock_query_dos_db.assert_has_calls(
+        calls=[
+            call(
+                connection=mock_connection,
+                query=(
+                    "SELECT id FROM symptomgroupsymptomdiscriminators WHERE symptomgroupid=%(SGID)s "
+                    "AND symptomdiscriminatorid=%(SDID)s;"
+                ),
+                query_vars={"SGID": 360, "SDID": 14167},
+            ),
+            call().close(),
+        ],
+    )
+
+
+@patch(f"{FILE_PATH}.query_dos_db")
+def test_validate_z_code_existss_does_not_exist(mock_query_dos_db: MagicMock):
+    # Arrange
+    mock_connection = MagicMock()
+    mock_query_dos_db.return_value.rowcount = 0
+    dos_service = MagicMock()
+    # Act
+    response = validate_z_code_exists(
+        connection=mock_connection,
+        dos_service=dos_service,
+        symptom_group_id=DOS_PALLIATIVE_CARE_SYMPTOM_GROUP,
+        symptom_discriminator_id=DOS_PALLIATIVE_CARE_SYMPTOM_DISCRIMINATOR,
+        z_code_alias="Palliative Care",
+    )
+    # Assert
+    assert False is response
+    mock_query_dos_db.assert_has_calls(
+        calls=[
+            call(
+                connection=mock_connection,
+                query=(
+                    "SELECT id FROM symptomgroupsymptomdiscriminators WHERE symptomgroupid=%(SGID)s "
+                    "AND symptomdiscriminatorid=%(SDID)s;"
+                ),
+                query_vars={"SGID": 360, "SDID": 14167},
+            ),
+            call().close(),
+        ],
+    )
