+++ test/integration/steps/conftest.py
@@ -4,6 +4,7 @@ from .functions.context import Context
 from .functions.generator import (
     set_up_blood_pressure_z_code_in_db,
     set_up_common_condition_service_types,
+    set_up_contraception_z_code_in_db,
     set_up_palliative_care_z_code_in_db,
 )
 
@@ -41,4 +42,5 @@ def pytest_sessionstart(session) -> None:
     """Called after the Session object has been created and before performing collection and entering the run test loop."""  # noqa: E501
     set_up_palliative_care_z_code_in_db()
     set_up_blood_pressure_z_code_in_db()
+    set_up_contraception_z_code_in_db()
     set_up_common_condition_service_types()
